Extract ClassInstrumentors from SandboxClassLoader.
